Develop Batch Jobs for Semi-Automatic NPC & Text ID Updates (Beta)
This commit provides Beta quality Batch Jobs for Updating NPC
Ids and Text Ids, as well as a utility job for generating mapping between POLUtils & Darkstar NPC Names These jobs, while much smarter and robust than my last attempt in the main repos pull request 134, still require some manual review and tweaking after they run, but are, in general, massive time savers.

Quick Summary:

GenerateNpcMappingFile.bat - Generates mappings between POLUtils & Darkstar NPC Names. Should be used when we are in a good state, and after manual fixes have been done. See readme for more details on limitations

NpcIdUpdate.bat - Attempts to auto update NPC ids after a version update. Accuracy varies mostly with the accuracy of the mapping file. See readme for more details on limitations.

TextIdUpdate.bat - Attempts to auto update Text Ids after a version update. Accuracy varies with the accuracy of the comments in the LUA file, as well as the state of Bad Chars in the POLUtils Dialog Table Dumps. See readme for more details on limitations. Entries that the job is not sure of get marked for manual review with a FIXME tag, and a summary of these instances is provided in the log after the job runs.

For more details on the jobs, how to use them, and how to improve them, see the README file :-) Both sources and binaries are provided, since they are small, use a different build process, and most people will be end users of it, not developers of it.

These jobs are currently beta quality at best, but should help us greatly after the update. As always, I am open to suggestions for improvements, and anyone who would like to help improve these is more than welcome to try.